task-04/pyproject.toml
[project]
name = "task-04"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Matteo Bongiorno", email = "pot8toking@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "openai>=2.7.2",
    "python-dotenv>=1.2.1",
]

[project.scripts]
task-04 = "task_04:main"

[build-system]
requires = ["uv_build>=0.9.7,<0.10.0"]
build-backend = "uv_build"
